implement self-checkout frontend for admins
