+++ self-checkout/src/components/ShoppingCart.jsx
@@ -1,20 +1,20 @@
 import React, { useState, useEffect, useCallback } from 'react';
-import CashSVG from '../assets/CashSVG';
-import VenmoSVG from '../assets/VenmoSVG';
+import PaymentPanel from './PaymentPanel';
+import AdminPanel from './AdminPanel';
 
 export default function ShoppingCart() {
-  const [items, setItems] = useState({});
-  const [currentInput, setCurrentInput] = useState('');
-  const [selectedMethod, setSelectedMethod] = useState('venmo')
-
-  const handleSubmit = (e) => {
-    e.preventDefault()
-    console.log(`Payment method: ${selectedMethod}`)
-    // Add confirm payment logic here
-  }
+  const [items, setItems] = useState({})
+  const [currentInput, setCurrentInput] = useState('')
+  const [mode, setMode] = useState('buyMode')
 
   const addItem = useCallback((itemName) => {
-    if (itemName in items) {
+    if (itemName === "buyMode") {
+      setMode("buyMode")
+    }
+    else if (itemName === "addMode") {
+      setMode("addMode")
+    }
+    else if (itemName in items) {
       setItems(prev => ({
         ...prev,
         [itemName]: {
@@ -33,7 +33,7 @@ export default function ShoppingCart() {
         }
       }));
     }
-  }, [items]);
+  }, [items, mode]);
 
   const increaseQuantity = useCallback((itemName) => {
     setItems(prev => ({
@@ -62,13 +62,6 @@ export default function ShoppingCart() {
     });
   }, []);
 
-  // const deleteItem = useCallback((itemName) => {
-  //   setItems(prev => {
-  //     const { [itemName]: _, ...rest } = prev;
-  //     return rest;
-  //   });
-  // }, []);
-
   const handleKeyDown = useCallback((event) => {
     if (event.key === 'Enter' && currentInput.trim() !== '') {
       addItem(currentInput.trim());
@@ -85,13 +78,13 @@ export default function ShoppingCart() {
     };
   }, [handleKeyDown]);
 
-  const total = Object.values(items).reduce((acc, item) => acc + item.price * item.quantity, 0);
-
   return (
     <div className="w-full px-3 h-[80vh] flex items-center gap-3">
       <div className="w-2/3 h-full rounded-lg flex flex-col items-center">
         <div className="w-full rounded-t-lg p-6 bg-primary text-center text-primary-content">
-          <h2 className="text-2xl font-bold">Shopping Cart</h2>
+          <h2 className="text-2xl font-bold">
+            {mode === "buyMode" ? "Shopping Cart" : "Admin Restock"}
+          </h2>
         </div>
         <div className="flex-grow w-full px-3 pt-3 bg-base-200 overflow-y-auto rounded-b-lg">
           {Object.keys(items).length === 0 ? (
@@ -115,73 +108,28 @@ export default function ShoppingCart() {
                   </button>
                   <button className="btn btn-square text-xl shadow-none">{items[item].quantity}</button>
                   <button className="btn btn-square shadow" onClick={() => increaseQuantity(item)}>
-                    <svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" stroke-width="1.5" stroke="currentColor" class="size-6">
-                      <path stroke-linecap="round" stroke-linejoin="round" d="M12 4.5v15m7.5-7.5h-15" />
+                    <svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" strokeWidth="1.5" stroke="currentColor" className="size-6">
+                      <path strokeLinecap="round" strokeLinejoin="round" d="M12 4.5v15m7.5-7.5h-15" />
                     </svg>
                   </button>
-                  {/* <button className="btn btn-square bg-red-400" onClick={() => deleteItem(item)}> */}
-                  {/*   <svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" strokeWidth={1.5} stroke="currentColor" className="size-6"> */}
-                  {/*     <path strokeLinecap="round" strokeLinejoin="round" d="m14.74 9-.346 9m-4.788 0L9.26 9m9.968-3.21c.342.052.682.107 1.022.166m-1.022-.165L18.16 19.673a2.25 2.25 0 0 1-2.244 2.077H8.084a2.25 2.25 0 0 1-2.244-2.077L4.772 5.79m14.456 0a48.108 48.108 0 0 0-3.478-.397m-12 .562c.34-.059.68-.114 1.022-.165m0 0a48.11 48.11 0 0 1 3.478-.397m7.5 0v-.916c0-1.18-.91-2.164-2.09-2.201a51.964 51.964 0 0 0-3.32 0c-1.18.037-2.09 1.022-2.09 2.201v.916m7.5 0a48.667 48.667 0 0 0-7.5 0" /> */}
-                  {/*   </svg> */}
-                  {/* </button> */}
                 </div>
               </div>
             ))
           )}
         </div>
       </div>
-
-      <div className="w-1/3 h-full rounded-lg flex flex-col items-center bg-base-200 p-3">
-        <div className="w-full flex justify-center h-[25vh] items-center text-center rounded-lg">
-          <span className="text-4xl font-bold text-success">Total:
-            <br />
-            <span className="text-7xl font-bold text-success">${total.toFixed(2)}</span>
-          </span>
-        </div>
-
-        <form onSubmit={handleSubmit} className="w-full flex-grow p-3 font-bold rounded-lg flex flex-col items-center justify-center">
-          <h2 className="text-2xl w-full text-center pb-3">Select Payment Method:</h2>
-          <div className="w-full items-center grid grid-cols-2 gap-3">
-            <label className={`cursor-pointer flex items-center text-center flex-col p-3 rounded-lg btn-ghost shadow ${selectedMethod === "venmo" ? "btn-active" : ""}`}>
-              <input
-                type="radio"
-                onChange={() => setSelectedMethod("venmo")}
-                value="venmo"
-                checked={selectedMethod === 'venmo'}
-                className="hidden"
-                id="venmo"
-              />
-              <VenmoSVG />
-              <span className="text-2xl font-bold mt-3">Venmo</span>
-            </label>
-
-            <label className={`cursor-pointer flex items-center text-center flex-col p-3 rounded-lg btn-ghost shadow ${selectedMethod === "cash" ? "btn-active" : ""}`}>
-              <input
-                type="radio"
-                onChange={() => setSelectedMethod("cash")}
-                value="cash"
-                checked={selectedMethod === 'cash'}
-                className="hidden"
-                id="cash"
-              />
-              <CashSVG />
-              <span className="text-2xl font-bold mt-3">Cash</span>
-            </label>
-          </div>
-          <button type="submit" className="btn btn-success text-2xl w-full p-6 my-3 h-auto rounded-lg">
-            Confirm Purchase
-          </button>
-        </form>
-
-        <div className="w-full p-6 font-bold rounded-lg flex flex-col items-center justify-center">
-          <h2 className="text-2xl w-full text-center">Coupon Code:</h2>
-          <div className="w-full flex-grow flex items-center justify-center">
-            <div className="flex items-center justify-center flex-col h-auto">
-              <h2 className="text-2xl w-full text-center">None</h2>
-            </div>
-          </div>
-        </div>
-      </div>
+      {mode === "buyMode" ? (
+        <PaymentPanel
+          items={items}
+          setItems={setItems}
+        />
+      ) : (
+        <AdminPanel
+          items={items}
+          setItems={setItems}
+        />
+      )
+      }
     </div>
   );
 }
